+++ libcxx/test/libcxx/algorithms/pstl.iterator-requirements.verify.cpp
@@ -142,6 +142,22 @@ void f(non_forward_iterator non_fwd,
   }
 
   {
+    (void)std::mismatch(pol, non_fwd, non_fwd, it); // expected-error@*:* {{static assertion failed: mismatch}}
+    (void)std::mismatch(pol, it, it, non_fwd);      // expected-error@*:* {{static assertion failed: mismatch}}
+
+    (void)std::mismatch(pol, non_fwd, non_fwd, it, it); // expected-error@*:* {{static assertion failed: mismatch}}
+    (void)std::mismatch(pol, it, it, non_fwd, non_fwd); // expected-error@*:* {{static assertion failed: mismatch}}
+
+    (void)std::mismatch(pol, non_fwd, non_fwd, it, pred); // expected-error@*:* {{static assertion failed: mismatch}}
+    (void)std::mismatch(pol, it, it, non_fwd, pred);      // expected-error@*:* {{static assertion failed: mismatch}}
+
+    (void)std::mismatch(
+        pol, non_fwd, non_fwd, it, it, pred); // expected-error@*:* {{static assertion failed: mismatch}}
+    (void)std::mismatch(
+        pol, it, it, non_fwd, non_fwd, pred); // expected-error@*:* {{static assertion failed: mismatch}}
+  }
+
+  {
     (void)std::move(pol, non_fwd, non_fwd, out); // expected-error@*:* {{static assertion failed: move}}
     (void)std::move(pol, it, it, non_fwd);       // expected-error@*:* {{static assertion failed: move}}
     (void)std::move(pol, it, it, non_output);    // expected-error@*:* {{static assertion failed: move}}